arrange city componnent structure
