+++ src/components/SideBar.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styles from "./Sidebar.module.css";
 import Logo from "./navigations/header/Logo";
-import AppNav from "./AppNav";
+import AppNav from "../components/appNav/AppNav";
 import Footer from "./navigations/footer/Footer";
 import { Outlet } from "react-router-dom";
