+++ src/components/button/Button.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import styles from '../Button.module.css'
+import styles from './Button.module.css'
 
 const Button = ({text, className, onClick, type}) => {
   return (
