+++ src/pages/AppLayout.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import AppNav from "../components/AppNav";
+import AppNav from "../components/appNav/AppNav";
 import SideBar from "../components/SideBar";
 import styles from "./AppLayout.module.css";
 import Map from "../components/map/Map";
